+++ source/ref_gl/r_mesh.c
@@ -672,7 +672,7 @@ void R_BuildTangentVectors( int numVertexes, vec4_t *xyzArray, vec4_t *normalsAr
 	vec3_t stackTVectorsArray[128];
 	vec3_t *tVectorsArray;
 
-	if( numVertexes > sizeof( stackTVectorsArray )/sizeof( stackTVectorsArray[0] ) )
+	if( (size_t)numVertexes > sizeof( stackTVectorsArray )/sizeof( stackTVectorsArray[0] ) )
 		tVectorsArray = R_Malloc( sizeof( vec3_t )*numVertexes );
 	else
 		tVectorsArray = stackTVectorsArray;
